bug fix
기존에 스플래쉬 스크린때 initial data를 불러오는 작업을 했으나 학교 서버문제로 오래걸리는 일이 종종 있어 분산시키고 lottie를 이용해 로딩화면 구성
키워드 등록 전 유효성 확인 등

+++ PushNotification/app/src/main/java/com/juhwan/anyang_yi/ui/notice/all/AllMainNoticeActivity.kt
@@ -7,6 +7,8 @@ import androidx.activity.viewModels
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.google.android.material.tabs.TabLayout
+import com.juhwan.anyang_yi.R
 import com.juhwan.anyang_yi.databinding.ActivityAllMainNoticeBinding
 
 class AllMainNoticeActivity : AppCompatActivity() {
@@ -14,7 +16,9 @@ class AllMainNoticeActivity : AppCompatActivity() {
     private lateinit var binding: ActivityAllMainNoticeBinding
     private val model: MainNoticeViewModel by viewModels()
     private lateinit var allMainNoticeAdapter: AllMainNoticeAdapter
-    private var page = 2
+    private var page = 1
+    private var bcIdx = "0"
+    private var isListEmpty = true
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -22,6 +26,8 @@ class AllMainNoticeActivity : AppCompatActivity() {
         setContentView(binding.root)
 
         initRecyclerView()
+        initTabLayout()
+        model.loadMainNotice(page, bcIdx)
 
         binding.ivBack.setOnClickListener {
             finish()
@@ -29,6 +35,7 @@ class AllMainNoticeActivity : AppCompatActivity() {
 
         model.getAll().observe(this, Observer {
             allMainNoticeAdapter.setList(it.resultList)
+            isListEmpty = false
             allMainNoticeAdapter.notifyItemRangeInserted((page - 1) * 15, 15)
         })
 
@@ -41,12 +48,11 @@ class AllMainNoticeActivity : AppCompatActivity() {
                 val itemTotalCount = recyclerView.adapter!!.itemCount-1
 
                 if (!binding.rvAllMainNotice.canScrollVertically(1) &&
-                    lastVisibleItemPosition == itemTotalCount
+                    lastVisibleItemPosition == itemTotalCount && !isListEmpty
                 ) {
                     allMainNoticeAdapter.deleteLoading()
-                    model.loadMainNotice(++page)
+                    model.loadMainNotice(++page, bcIdx)
                 }
-
             }
         })
     }
@@ -55,5 +61,37 @@ class AllMainNoticeActivity : AppCompatActivity() {
         binding.rvAllMainNotice.layoutManager = LinearLayoutManager(this)
         allMainNoticeAdapter = AllMainNoticeAdapter()
         binding.rvAllMainNotice.adapter = allMainNoticeAdapter
+    }
+
+    private fun initTabLayout() {
+        binding.tabLayoutMainNoticeMenu.tabTextColors = resources.getColorStateList(R.color.tab_icon, null)
+        binding.tabLayoutMainNoticeMenu.addOnTabSelectedListener(object: TabLayout.OnTabSelectedListener{
+            override fun onTabReselected(tab: TabLayout.Tab?) {
+
+            }
+
+            override fun onTabUnselected(tab: TabLayout.Tab?) {
+
+            }
+
+            override fun onTabSelected(tab: TabLayout.Tab?) {
+                // bcIdx
+                // 전체(0), 대학교(20), 학사(80), 학사(21), 취업정보(23), 입찰채용(24)
+                when(tab!!.position){
+                    0 -> bcIdx = "0"
+                    1 -> bcIdx = "20"
+                    2 -> bcIdx = "80"
+                    3 -> bcIdx = "21"
+                    4 -> bcIdx = "23"
+                    5 -> bcIdx = "24"
+                }
+
+                page = 1
+                allMainNoticeAdapter.resetList(bcIdx)
+                isListEmpty = true
+                allMainNoticeAdapter.notifyDataSetChanged()
+                model.loadMainNotice(page, bcIdx)
+            }
+        })
     }
 }
